Extract WinUI-independent UI core
