+++ gorilla-ui/src/Gorilla.UI.App/Services/GorillaUiServices.cs
@@ -1,4 +1,5 @@
 using Gorilla.UI.Client;
+using Gorilla.UI.Core;
 
 namespace Gorilla.UI.App.Services;
