+++ gorilla-ui/src/Gorilla.UI.App/Services/JsonFileOptionalInstallsCacheStore.cs
@@ -1,18 +1,8 @@
 using System.Text.Json;
+using Gorilla.UI.Client;
+using Gorilla.UI.Core;
 
-namespace Gorilla.UI.Client;
-
-public sealed record OptionalInstallsCacheDocument(
-    DateTimeOffset CachedAtUtc,
-    IReadOnlyList<OptionalInstallItem> Items
-);
-
-public interface IOptionalInstallsCacheStore
-{
-    Task<OptionalInstallsCacheDocument?> LoadAsync(CancellationToken cancellationToken);
-
-    Task SaveAsync(OptionalInstallsCacheDocument document, CancellationToken cancellationToken);
-}
+namespace Gorilla.UI.App.Services;
 
 public sealed class JsonFileOptionalInstallsCacheStore : IOptionalInstallsCacheStore
 {
@@ -73,34 +63,5 @@ public sealed class JsonFileOptionalInstallsCacheStore : IOptionalInstallsCacheS
 
         await using var stream = File.Create(_cacheFilePath);
         await JsonSerializer.SerializeAsync(stream, document, ProtocolJson.Options, cancellationToken);
-    }
-}
-
-public sealed class OptionalInstallsCacheCoordinator
-{
-    private readonly IGorillaServiceClient _client;
-    private readonly IOptionalInstallsCacheStore _cacheStore;
-
-    public OptionalInstallsCacheCoordinator(IGorillaServiceClient client, IOptionalInstallsCacheStore cacheStore)
-    {
-        _client = client;
-        _cacheStore = cacheStore;
-    }
-
-    public Task<OptionalInstallsCacheDocument?> LoadCachedAsync(CancellationToken cancellationToken)
-    {
-        return _cacheStore.LoadAsync(cancellationToken);
-    }
-
-    public async Task<OptionalInstallsCacheDocument> RefreshAsync(CancellationToken cancellationToken)
-    {
-        var items = await _client.ListOptionalInstallsAsync(cancellationToken);
-        var document = new OptionalInstallsCacheDocument(
-            CachedAtUtc: DateTimeOffset.UtcNow,
-            Items: items
-        );
-
-        await _cacheStore.SaveAsync(document, cancellationToken);
-        return document;
     }
 }
