+++ gorilla-ui/src/Gorilla.UI.App/Views/HomePage.xaml.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Gorilla.UI.App.ViewModels;
+using Gorilla.UI.Core.ViewModels;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
