+++ gorilla-ui/src/Gorilla.UI.Core/Gorilla.UI.Core.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../Gorilla.UI.Client/Gorilla.UI.Client.csproj" />
+  </ItemGroup>
+</Project>
