+++ gorilla-ui/src/Gorilla.UI.Core/Services/OperationTracker.cs
@@ -1,9 +1,6 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using Gorilla.UI.Client;
 
-namespace Gorilla.UI.App.Services;
+namespace Gorilla.UI.Core.Services;
 
 public sealed class OperationTracker
 {
